+++ packages/gov-ui-kit/CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 
 -   Implement `VoteProposalDataListItem` with Structure and Skeleton module
+-   Add `useBlockExplorer` hook to generate block explorer links
 
 ### Changed